+++ packages/web-app/src/modules/error-views/FirewallErrorContainer.ts
@@ -1,0 +1,17 @@
+import { connect } from '../../connect'
+import { RootStore } from '../../Store'
+import { getSanitizedHTML } from '../../utils'
+import { AntiVirusFirewallErrorPage } from './components/AntiVirusFirewallErrorPage'
+
+const mapStoreToProps = (store: RootStore): any => {
+  return {
+    errorType: store.zendesk.errorType,
+    article: store.zendesk.helpCenterArticle ? getSanitizedHTML(store.zendesk.helpCenterArticle) : undefined,
+    fallthrough: store.ui.hasViewedFirewallErrorPage,
+    loading: store.zendesk.loadingArticle,
+    loadArticle: () => store.zendesk.loadFirewallArticle(),
+    onCloseClicked: () => store.ui.hideModal(true),
+  }
+}
+
+export const FirewallErrorContainer = connect(mapStoreToProps, AntiVirusFirewallErrorPage)
