+++ packages/web-app/src/modules/error-views/index.ts
@@ -5,3 +5,4 @@ export * from './NetworkErrorContainer'
 export * from './NotCompatibleErrorContainer'
 export * from './SpecificAntiVirusErrorContainer'
 export * from './UnknownErrorContainer'
+export * from './FirewallErrorContainer'
