+++ packages/web-app/src/modules/salad-bowl/models/ErrorCategory.ts
@@ -2,6 +2,7 @@ export enum ErrorCategory {
   Unknown = 'unknown',
   Driver = 'driver',
   AntiVirus = 'antiVirus',
+  Firewall = 'firewall',
   Network = 'network',
   Silent = 'silent',
 }
